adds basic support for event frames
