+++ PIconnect/PIAF.py
@@ -34,10 +34,10 @@ except ImportError:
 # pragma pylint: enable=unused-import, redefined-builtin
 from warnings import warn
 
-from PIconnect._operators import OPERATORS, add_operators
-from PIconnect.AFSDK import AF
-from PIconnect.PIData import PISeries, PISeriesContainer
 from PIconnect._utils import classproperty
+from PIconnect.AFSDK import AF
+from PIconnect.PIAFBase import PIAFBaseElement
+from PIconnect.PIConsts import EventFrameSearchMode
 
 _NOTHING = object()
 
@@ -48,7 +48,7 @@ class PIAFDatabase(object):
     Context manager for connections to the PI Asset Framework database.
     """
 
-    version = "0.1.2"
+    version = "0.2.0"
 
     _servers = _NOTHING
     _default_server = _NOTHING
@@ -147,22 +147,38 @@ class PIAFDatabase(object):
         """Return a descendant of the database from an exact path."""
         return PIAFElement(self.database.Elements.get_Item(path))
 
+    def event_frames(
+        self,
+        start_time=None,
+        start_index=0,
+        max_count=1000,
+        search_mode=EventFrameSearchMode.STARTING_AFTER,
+        search_full_hierarchy=False,
+    ):
+        if not start_time:
+            start_time = AF.Time.AFTime.Now
+        return {
+            frame.Name: PIAFEventFrame(frame)
+            for frame in AF.EventFrame.AFEventFrame.FindEventFrames(
+                self.database,
+                None,
+                start_time,
+                start_index,
+                max_count,
+                search_mode,
+                None,
+                None,
+                None,
+                None,
+                search_full_hierarchy,
+            )
+        }
 
-class PIAFElement(object):
+
+class PIAFElement(PIAFBaseElement):
     """Container for PI AF elements in the database."""
 
     version = "0.1.0"
-
-    def __init__(self, element):
-        self.element = element
-
-    def __repr__(self):
-        return "%s(%s)" % (self.__class__.__name__, self.name)
-
-    @property
-    def name(self):
-        """Return the name of the current element."""
-        return self.element.Name
 
     @property
     def parent(self):
@@ -180,132 +196,19 @@ class PIAFElement(object):
         """Return a descendant of the current element from an exact path."""
         return self.__class__(self.element.Elements.get_Item(path))
 
-    @property
-    def attributes(self):
-        """Return a dictionary of the attributes of the current element."""
-        return {a.Name: PIAFAttribute(self, a) for a in self.element.Attributes}
 
-
-@add_operators(
-    operators=OPERATORS,
-    members=["_current_value", "interpolated_values"],
-    newclassname="VirtualPIAFAttribute",
-    attributes=["element", "attribute"],
-)
-class PIAFAttribute(PISeriesContainer):
-    """Container for attributes of PI AF elements in the database."""
+class PIAFEventFrame(PIAFBaseElement):
+    """Container for PI AF Event Frames in the database."""
 
     version = "0.1.0"
 
-    def __init__(self, element, attribute):
-        super().__init__()
-        self.element = element
-        self.attribute = attribute
-
-    def __repr__(self):
-        return "%s(%s, %s; Current Value: %s %s)" % (
-            self.__class__.__name__,
-            self.name,
-            self.description,
-            self.current_value,
-            self.units_of_measurement,
-        )
-
     @property
-    def name(self):
-        """Return the name of the current attribute."""
-        return self.attribute.Name
+    def event_frame(self):
+        return self.element
 
     @property
     def parent(self):
-        """Return the parent attribute of the current attribute, or None if it has none."""
-        if not self.attribute.Parent:
+        """Return the parent element of the current element, or None if it has none."""
+        if not self.element.Parent:
             return None
-        return self.__class__(self.element, self.attribute.Parent)
-
-    @property
-    def children(self):
-        """Return a dictionary of the direct child attributes of the current attribute."""
-        return {
-            a.Name: self.__class__(self.element, a) for a in self.attribute.Attributes
-        }
-
-    @property
-    def description(self):
-        """Return the description of the PI Point."""
-        return self.attribute.Description
-
-    @property
-    def last_update(self):
-        """Return the time at which the current_value was last updated."""
-        return PISeries.timestamp_to_index(self.attribute.GetValue().Timestamp.UtcTime)
-
-    @property
-    def units_of_measurement(self):
-        """Return the units of measurement in which values for this element are reported."""
-        return self.attribute.DefaultUOM
-
-    def _current_value(self):
-        return self.attribute.GetValue().Value
-
-    def _update_value(self, value, update_mode, buffer_mode):
-        return self.attribute.Data.UpdateValue(
-            value,
-            update_mode,
-            buffer_mode,
-        )
-
-    def _recorded_values(self, time_range, boundary_type, filter_expression):
-        include_filtered_values = False
-        return self.attribute.Data.RecordedValues(
-            time_range,
-            boundary_type,
-            self.attribute.DefaultUOM,
-            filter_expression,
-            include_filtered_values,
-        )
-
-    def _interpolated_values(self, time_range, interval, filter_expression):
-        """Internal function to actually query the pi point"""
-        include_filtered_values = False
-        return self.attribute.Data.InterpolatedValues(
-            time_range,
-            interval,
-            self.attribute.DefaultUOM,
-            filter_expression,
-            include_filtered_values,
-        )
-
-    def _summary(self, time_range, summary_types, calculation_basis, time_type):
-        return self.attribute.Data.Summary(
-            time_range, summary_types, calculation_basis, time_type
-        )
-
-    def _summaries(
-        self, time_range, interval, summary_types, calculation_basis, time_type
-    ):
-        return self.attribute.Data.Summaries(
-            time_range, interval, summary_types, calculation_basis, time_type
-        )
-
-    def _filtered_summaries(
-        self,
-        time_range,
-        interval,
-        filter_expression,
-        summary_types,
-        calculation_basis,
-        filter_evaluation,
-        filter_interval,
-        time_type,
-    ):
-        return self.attribute.Data.FilteredSummaries(
-            time_range,
-            interval,
-            filter_expression,
-            summary_types,
-            calculation_basis,
-            filter_evaluation,
-            filter_interval,
-            time_type,
-        )
+        return self.__class__(self.element.Parent)
