+++ PIconnect/PIConsts.py
@@ -152,6 +152,39 @@ class TimestampCalculation(IntEnum):
     MOST_RECENT_TIME = 2
 
 
+class EventFrameSearchMode(IntEnum):
+    """EventFrameSearchMode
+
+    EventFrameSearchMode defines the interpretation and direction from the start time
+    when searching for event frames.
+
+    Detailed information is available at https://techsupport.osisoft.com/Documentation/PI-AF-SDK/html/T_OSIsoft_AF_EventFrame_AFEventFrameSearchMode.htm,
+    including a graphical display of event frames that are returned for a given search
+    mode.
+    """
+
+    #: Uninitialized
+    NONE = 0
+    #: Backward from start time, also known as starting before
+    BACKWARD_FROM_START_TIME = 1
+    STARTING_BEFORE = 1
+    #: Forward from start time, also known as starting after
+    FORWARD_FROM_START_TIME = 2
+    STARTING_AFTER = 2
+    #: Backward from end time, also known as ending before
+    BACKWARD_FROM_END_TIME = 3
+    ENDING_BEFORE = 3
+    #: Forward from end time, also known as ending after
+    FORWARD_FROM_END_TIME = 4
+    ENDING_AFTER = 4
+    #: Backward in progress, also known as starting before and in progress
+    BACKWARD_IN_PROGRESS = 5
+    STARTING_BEFORE_IN_PROGRESS = 5
+    #: Forward in progress, also known as starting after and in progress
+    FORWARD_IN_PROGRESS = 6
+    STARTING_AFTER_IN_PROGRESS = 6
+
+
 def get_enumerated_value(enumeration, value, default):
     if not value:
         return default
