+++ PIconnect/__init__.py
@@ -2,7 +2,7 @@
     Connector to the OSISoft PI and PI-AF databases.
 """
 # pragma pylint: disable=unused-import
-from PIconnect.AFSDK import AF
+from PIconnect.AFSDK import AF, AF_SDK_VERSION
 from PIconnect.config import PIConfig
 from PIconnect.PI import PIServer
 from PIconnect.PIAF import PIAFDatabase
